[FIX] web: domain field: detect expressions in sub trees
If allowExpressions = false, the domain field is supposed to detect when
expressions are used in the domain and notify the user that the domain
is invalid. The current version of domainContainsExpressions do not
allow detection of expressions in sub domains used in conjunction with
the any operator. We fix that.

closes odoo/odoo#212957

Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>