[FIX] payment_demo: fix demo express checkout
Versions
--------
- 18.0+

Steps
-----
1. Enable Demo express checkout;
2. pay with Demo Express checkout in eCommerce.

Issue
-----
Validation error: No delivery method is available.

Cause
-----
Commit bf8d08cd22b5 changed the express checkout delivery method RPC
API. Instead of an array, `availableCarriers` is now an object with a
`delivery_methods` attribute.

Because it checks the length of an object, which is `undefined`, Demo
express checkout incorrectly believes no delivery methods are available.

Solution
--------
Check the length of `delivery_methods` instead of `availableCarriers`.

closes odoo/odoo#213240

Signed-off-by: Valeriya Chuprina (vchu) <vchu@odoo.com>