[REV] l10n_us: add US tax report variant
This reverts commit 6d0d0e58b9f2c4ee2519e8e960e156c7e7be9841.

The deployment of the reverted commit revealed an error: l10n_us_reports does not depend on l10n_us_account (only l10n_us), so it can't override the US tax report in data and crashes. We sadly cannot fix that in stable ; we'll do it in master.

closes odoo/odoo#213363

Related: odoo/enterprise#87195
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>