[FIX] mail: show display_name for contact without name and email
Steps to reproduce
===============
1. Create a contact of type invoice address without name and email.
2. Go to any app with chatter.
3. Add this user to the recipient

----> Only the blue tick will be visible (recipient name will be empty)

Along with this, if there is any follower without a name and email,
false was shown in the recipient list above the composer and in the
follower list (which can be opened by pressing the down arrow next to
the recipients).

After this commit, we will use the display_name as a fallback to show in
those places.

Note: Added/adapted tests to ensure the correct behaviour. Also some
mocked models were missing the definition for computing correct display
name. Thus, compute methods are overriden in those mocked models.

Task-4812554

closes odoo/odoo#213545

Related: odoo/enterprise#91003
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>